pyproject.toml
[tool.poetry]
name = "actions-fargate"
version = "0.1.0"
description = "Farget service management utility"
authors = ["Epsy <engineering@epsyhealth.com>"]
packages = [
    { include = "fargate" }
]


[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
boto3 = "^1.16.14"
python-box = "^5.2.0"
arrow = "^0.17.0"

[tool.poetry.scripts]
fargate = 'fargate.cli:run'

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
